[WCNCR00240986] connac3: Support Falcon USB
[Description]
Support Falcon USB

[Unit Test]
Falcon USB FPGA M2M with Bellwether
  - ping stress pass
  - iperf UL stress pass (> 6 hours)
  - iperf DL stress pass (> 6 hours)

MTK-Commit-Id: 27d67f2717cdce26b952509c1b7c233e531bd895

CR-Id: WCNCR00240986
Feature: connac3
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/5595611
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_mspautosanity

+++ common/wlan_lib.c
@@ -12898,8 +12898,7 @@ void wlanFinishCollectingLinkQuality(struct GLUE_INFO *prGlueInfo)
 			  prLinkQualityInfo->u8LastTxFailCount) : 0;
 	if (u8TxTotalCntDif >= u8TxFailCntDif)
 		prLinkQualityInfo->u4CurTxPer = (u8TxTotalCntDif == 0) ? 0 :
-			((uint32_t)(u8TxFailCntDif * 100) /
-				(uint32_t)u8TxTotalCntDif);
+			((u8TxFailCntDif * 100) / u8TxTotalCntDif);
 	else
 		prLinkQualityInfo->u4CurTxPer = 0;
